Refactor: comment & input func refactor
