+++ src/pages/byeonghyeon/Main/component/feeds/feeds.scss
@@ -1,10 +1,11 @@
 .feeds {
   width: 58%;
-  background: #fff;
+  background: #fafafa;
   .feed {
     border: 1px solid #dee2e6;
     border-radius: 2px;
     height: fit-content;
+    margin-bottom: 20px;
     .profile-container {
       display: flex;
       justify-content: space-between;
