+++ src/pages/byeonghyeon/data/FeedData.json
@@ -2,19 +2,19 @@
   {
     "id": 1,
     "userName": "choi_95",
-    "profileImg": "/images/byeonghyeon/IMG_1405.JPG",
+    "profileImg": "/images/byeonghyeon/IMG_1402.JPG",
     "feedImg": "/images/byeonghyeon/IMG_1405.JPG"
   },
   {
     "id": 2,
     "userName": "orange_96",
-    "profileImg": "/images/byeonghyeon/images/IMG_1.JPG",
+    "profileImg": "/images/byeonghyeon/IMG_1.JPG",
     "feedImg": "/images/byeonghyeon/IMG_6941.JPG"
   },
   {
     "id": 3,
     "userName": "apple_90",
     "profileImg": "/images/byeonghyeon/IMG_2.JPG",
-    "feedImg": "/images/byoenghyeon/IMG_6944.JPG"
+    "feedImg": "/images/byeonghyeon/IMG_6944.JPG"
   }
 ]
